LayoutTests/imported/w3c/web-platform-tests/import-maps/data-url-specifiers.sub-expected.txt

CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
Blocked access to external URL https://www1.localhost:9443/import-maps/resources/log.js?pipe=sub&name=cross-origin-foo
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
Blocked access to external URL https://www1.localhost:9443/import-maps/resources/log.js?pipe=sub&name=cross-origin-foo
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
Blocked access to external URL https://www1.localhost:9443/import-maps/resources/log.js?pipe=sub&name=cross-origin-foo
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: TypeError: Module name, 'data:text/javascript,log.push('data:to-bare')' does not resolve to a valid URL.
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare
CONSOLE MESSAGE: value in specifier map cannot be parsed as URL bare


PASS data:text/javascript,log.push('data:foo'): <script src type=module>
PASS data:text/javascript,log.push('data:foo'): <script src type=text/javascript>
PASS data:text/javascript,log.push('data:foo'): static import
PASS data:text/javascript,log.push('data:foo'): dynamic import (from module)
PASS data:text/javascript,log.push('data:foo'): dynamic import (from text/javascript)
PASS data:text/javascript,log.push('data:cross-origin-foo'): <script src type=module>
PASS data:text/javascript,log.push('data:cross-origin-foo'): <script src type=text/javascript>
FAIL data:text/javascript,log.push('data:cross-origin-foo'): static import assert_unreached: script's error event shouldn't be fired Reached unreachable code
FAIL data:text/javascript,log.push('data:cross-origin-foo'): dynamic import (from module) assert_unreached: dynamic import promise shouldn't be rejected Reached unreachable code
FAIL data:text/javascript,log.push('data:cross-origin-foo'): dynamic import (from text/javascript) assert_unreached: dynamic import promise shouldn't be rejected Reached unreachable code
PASS data:text/javascript,log.push('data:to-data'): <script src type=module>
PASS data:text/javascript,log.push('data:to-data'): <script src type=text/javascript>
PASS data:text/javascript,log.push('data:to-data'): static import
PASS data:text/javascript,log.push('data:to-data'): dynamic import (from module)
PASS data:text/javascript,log.push('data:to-data'): dynamic import (from text/javascript)
PASS data:text/javascript,log.push('data:to-bare'): <script src type=module>
PASS data:text/javascript,log.push('data:to-bare'): <script src type=text/javascript>
PASS data:text/javascript,log.push('data:to-bare'): static import
PASS data:text/javascript,log.push('data:to-bare'): dynamic import (from module)
PASS data:text/javascript,log.push('data:to-bare'): dynamic import (from text/javascript)
